Fixed error with JS function called on all views
