+++ app/javascript/packs/application.js
@@ -1,8 +1,4 @@
 // import "bootstrap";
 import { subnav } from '../components/subnav';
-import { formAmount } from '../components/formamount'
 
 subnav();
-formAmount();
-
-
